Fixes after moving files from blmc_robots
Adjust namespaces, include paths, etc.
