+++ include/robot_fingers/two_joint_driver.hpp
@@ -7,7 +7,7 @@
 
 #pragma once
 
-#include <blmc_robots/n_joint_blmc_robot_driver.hpp>
+#include <robot_fingers/n_joint_blmc_robot_driver.hpp>
 
 namespace robot_fingers
 {
@@ -16,7 +16,7 @@ namespace robot_fingers
  *
  * Driver for a double BLMC joint.  Mostly intended for testing purposes.
  */
-class TwoJointDriver : public blmc_robots::SimpleNJointBlmcRobotDriver<2, 1>
+class TwoJointDriver : public SimpleNJointBlmcRobotDriver<2, 1>
 {
 public:
     TwoJointDriver(const Config &config)
@@ -26,7 +26,7 @@ public:
 
 private:
     TwoJointDriver(const MotorBoards &motor_boards, const Config &config)
-        : blmc_robots::SimpleNJointBlmcRobotDriver<2, 1>(
+        : SimpleNJointBlmcRobotDriver<2, 1>(
               motor_boards,
               create_motors(motor_boards),
               {
