+++ srcpy/py_real_finger.cpp
@@ -25,7 +25,6 @@
 
 using namespace pybind11::literals;
 using namespace robot_fingers;
-using namespace blmc_robots;
 
 PYBIND11_MODULE(py_real_finger, m)
 {
