+++ srcpy/py_trifinger.cpp
@@ -26,7 +26,6 @@
 
 using namespace pybind11::literals;
 using namespace robot_fingers;
-using namespace blmc_robots;
 
 PYBIND11_MODULE(py_trifinger, m)
 {
