+++ test/test_clamp.cpp
@@ -6,12 +6,12 @@
 #include <gtest/gtest.h>
 #include <Eigen/Eigen>
 
-#include <blmc_robots/clamp.hpp>
+#include <robot_fingers/clamp.hpp>
 
 TEST(TestClamp, clamp_within)
 {
     Eigen::Vector3d vec(1, 2, 3);
-    Eigen::Vector3d clamped = blmc_robots::clamp(vec, 0, 4);
+    Eigen::Vector3d clamped = robot_fingers::clamp(vec, 0, 4);
     Eigen::Vector3d expected(1, 2, 3);
     ASSERT_EQ(expected, clamped);
 }
@@ -19,7 +19,7 @@ TEST(TestClamp, clamp_within)
 TEST(TestClamp, clamp_below)
 {
     Eigen::Vector3d vec(1, 2, 3);
-    Eigen::Vector3d clamped = blmc_robots::clamp(vec, 2, 4);
+    Eigen::Vector3d clamped = robot_fingers::clamp(vec, 2, 4);
     Eigen::Vector3d expected(2, 2, 3);
     ASSERT_EQ(expected, clamped);
 }
@@ -27,7 +27,7 @@ TEST(TestClamp, clamp_below)
 TEST(TestClamp, clamp_above)
 {
     Eigen::Vector4d vec(1, 2, 3, 4);
-    Eigen::Vector4d clamped = blmc_robots::clamp(vec, 1, 3);
+    Eigen::Vector4d clamped = robot_fingers::clamp(vec, 1, 3);
     Eigen::Vector4d expected(1, 2, 3, 3);
     ASSERT_EQ(expected, clamped);
 }
@@ -35,7 +35,13 @@ TEST(TestClamp, clamp_above)
 TEST(TestClamp, clamp_below_and_above)
 {
     Eigen::Vector4d vec(1, 2, 3, 4);
-    Eigen::Vector4d clamped = blmc_robots::clamp(vec, 2, 3);
+    Eigen::Vector4d clamped = robot_fingers::clamp(vec, 2, 3);
     Eigen::Vector4d expected(2, 2, 3, 3);
     ASSERT_EQ(expected, clamped);
+}
+
+int main(int argc, char **argv)
+{
+    ::testing::InitGoogleTest(&argc, argv);
+    return RUN_ALL_TESTS();
 }
